pbench_run duplicate removal
